control

+++ f1-balance-bike-control/User/Inc/data_proc.h
@@ -4,5 +4,5 @@
 #include "wireless.h"
 
 void gimbal_angle_calc(float *angle_buttom_offset,float *angle_top_offset);
-void motor_speed_calc(int *motor_left_speed,int *motor_right_speed);
+void motor_speed_calc(int16_t *motor_left_speed,int16_t *motor_right_speed);
 void data_packaing(CommandPacket *command);
