+++ f4-balance-bike/User/include/wireless.h
@@ -20,7 +20,7 @@ typedef struct {
     CommandType type;
     union {
         struct {
-            int speed[2];
+            int16_t speed[2];
         } move;
         struct {
             float angle[2];
